move the trait into a separate module
