+++ src/index.rs
@@ -1,29 +1,8 @@
 use geo::Polygon;
-use geoarrow::{
-    array::PolygonArray, scalar::Polygon as GeoArrowPolygon, trait_::ArrayAccessor,
-    trait_::NativeScalar,
-};
-use pyo3::prelude::*;
-use rstar::{primitives::CachedEnvelope, RTree, RTreeObject};
+use geoarrow::array::PolygonArray;
+use rstar::{primitives::CachedEnvelope, RTree};
 
-trait RStarRTree<T: RTreeObject> {
-    fn create_rstar_rtree(self) -> RTree<T>;
-}
-
-impl RStarRTree<CachedEnvelope<Polygon>> for PolygonArray {
-    fn create_rstar_rtree(self) -> RTree<CachedEnvelope<Polygon>> {
-        let cells: Vec<_> = self
-            .iter()
-            .flatten()
-            .map(|cell| {
-                let geom: Polygon = cell.to_geo();
-                CachedEnvelope::<Polygon>::new(geom.into())
-            })
-            .collect();
-
-        return RTree::bulk_load_with_params(cells);
-    }
-}
+use super::trait_::RStarRTree;
 
 pub struct Index {
     rtree: RTree<CachedEnvelope<Polygon>>,
